separate file for offer button
