+++ src/nft/components/details/detailsV2/BuyButton.tsx
@@ -1,27 +1,13 @@
 import { Trans } from '@lingui/macro'
 import { formatNumber } from '@uniswap/conedison/format'
-import { ButtonGray, ButtonPrimary } from 'components/Button'
+import { ButtonPrimary } from 'components/Button'
 import Loader from 'components/Icons/LoadingSpinner'
-import { HandHoldingDollarIcon } from 'nft/components/icons'
 import { useFetchSingleAsset } from 'nft/hooks/useFetchAssets'
 import { GenieAsset } from 'nft/types'
-import styled, { css } from 'styled-components/macro'
+import styled from 'styled-components/macro'
 
-const ButtonStyles = css`
-  width: min-content;
-  flex-shrink: 0;
-  border-radius: 16px;
-`
-
-const MakeOfferButtonSmall = styled(ButtonPrimary)`
-  padding: 16px;
-  ${ButtonStyles}
-`
-
-const MakeOfferButtonLarge = styled(ButtonGray)`
-  white-space: nowrap;
-  ${ButtonStyles}
-`
+import { OfferButton } from './OfferButton'
+import { ButtonStyles } from './shared'
 
 const StyledBuyButton = styled(ButtonPrimary)`
   display: flex;
@@ -43,11 +29,7 @@ export const BuyButton = ({ asset, onDataPage }: { asset: GenieAsset; onDataPage
   const price = asset.sellorders?.[0]?.price.value
 
   if (!price) {
-    return (
-      <MakeOfferButtonLarge>
-        <Trans>Make an offer</Trans>
-      </MakeOfferButtonLarge>
-    )
+    return <OfferButton />
   }
 
   return (
@@ -65,11 +47,7 @@ export const BuyButton = ({ asset, onDataPage }: { asset: GenieAsset; onDataPage
           </>
         )}
       </StyledBuyButton>
-      {onDataPage && (
-        <MakeOfferButtonSmall>
-          <HandHoldingDollarIcon />
-        </MakeOfferButtonSmall>
-      )}
+      {onDataPage && <OfferButton smallVersion />}
     </>
   )
 }
